Add support for custom frame
Add a method -(id)initWithFrame:InScrollView: in ODRefreshControl to
allow users of the class to provide a custom frame. Currently,
ODRefreshControl is assuming that the parent UITableView has vertical
UITableViewCells, which may not be the case in apps like Pulse-news,
where the UITableViewCells are layed out horizontally on the
UITableView. Add the aforementioned method to allow assigning custom
frame for the control so that it can animate properly in a horizontal
UITableView.